change components
